requirements.txt
# ----------------------------------------------------------
# Author: Nandan Kumar
# Date: 11/18/2025
# Assignment-11: Calculation Model, Pydantic Validation,
#                Factory Pattern & Docker Testing
# File: requirements.txt
# ----------------------------------------------------------

# ----------------------------------------------------------
# 1. Core Framework
# ----------------------------------------------------------
fastapi==0.115.4
uvicorn[standard]==0.32.0
starlette==0.41.2
Jinja2==3.1.4
MarkupSafe==3.0.2

# ----------------------------------------------------------
# 2. Validation & Typing
# ----------------------------------------------------------
pydantic==2.9.2
pydantic_core==2.23.4
pydantic-settings==2.7.1
annotated-types==0.7.0
typing_extensions==4.12.2
email_validator==2.2.0

# ----------------------------------------------------------
# 3. Database & ORM
# ----------------------------------------------------------
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
greenlet==3.1.1
python-dotenv==1.0.1

# ----------------------------------------------------------
# 4. Security & Authentication
# ----------------------------------------------------------
bcrypt==4.2.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
cryptography==44.0.0
rsa==4.9
ecdsa==0.19.0
pyasn1==0.6.1
cffi==1.17.1
pycparser==2.22
PyJWT==2.9.0

# ----------------------------------------------------------
# 5. HTTP & Networking
# ----------------------------------------------------------
requests==2.32.3
httpx==0.27.2
anyio==4.6.2.post1
sniffio==1.3.1
urllib3==2.2.3
charset-normalizer==3.4.0
certifi==2024.8.30
tenacity==9.0.0
python-multipart==0.0.20

# ----------------------------------------------------------
# 6. Testing & Coverage
# ----------------------------------------------------------
pytest==8.3.3
pytest-cov==6.0.0
pytest-asyncio==0.24.0
coverage==7.6.4
Faker==33.3.0

# ----------------------------------------------------------
# 7. E2E Testing (Optional but allowed)
# ----------------------------------------------------------
playwright==1.48.0
pyee==12.0.0

# ----------------------------------------------------------
# 8. Linting & Code Quality
# ----------------------------------------------------------
pylint==3.3.1
isort==5.13.2
astroid==3.3.5
mccabe==0.7.0
dill==0.3.9
platformdirs==4.3.6
packaging==24.2
tomli==2.0.2
tomlkit==0.13.2
six==1.17.0

# ----------------------------------------------------------
# 9. Utilities
# ----------------------------------------------------------
python-dateutil==2.9.0.post0
click==8.1.7
